refactor: relocate array udfs
