+++ hybridse/src/udf/default_defs/containers.h
@@ -18,6 +18,7 @@
 #define HYBRIDSE_SRC_UDF_DEFAULT_DEFS_CONTAINERS_H_
 
 #include <string>
+#include <tuple>
 #include <utility>
 
 #include "udf/containers.h"
@@ -27,6 +28,8 @@
 namespace hybridse {
 namespace udf {
 namespace container {
+
+using hybridse::udf::ArrayRef;
 
 // Top N Value wrapper over *CateWhere
 // Base template class for 'top_n_value_*_cate_where' udafs
@@ -96,6 +99,37 @@ struct TopNValueImpl {
             ptr->~ContainerT();
         }
     };
+};
+
+// ============================================ //
+//     Array UDF containers
+// ============================================ //
+template <typename T>
+struct ArrayContains {
+    using ParamType = typename DataTypeTrait<T>::CCallArgType;
+
+    // udf registry types
+    using Args = std::tuple<ArrayRef<T>, T>;
+
+    // type binding, udf registry type -> function param type
+    // - bool/intxx/float/double -> bool/intxx/float/double
+    // - Timestamp/Date/StringRef -> Timestamp*/Date*/StringRef*
+    bool operator()(ArrayRef<ParamType>* arr, ParamType v) {
+        // NOTE: array_contains([null], null) returns null
+        // this might not expected
+        for (uint64_t i = 0; i < arr->size; ++i) {
+            if constexpr (std::is_pointer_v<ParamType>) {
+                if (!arr->nullables[i] && *arr->raw[i] == *v) {
+                    return true;
+                }
+            } else {
+                if (!arr->nullables[i] && arr->raw[i] == v) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
 };
 
 }  // namespace container
