+++ hybridse/src/udf/default_udf_library.h
@@ -48,6 +48,9 @@ class DefaultUdfLibrary : public UdfLibrary {
     void initMaxByCateUdaFs();
     void InitAvgByCateUdafs();
     void InitFeatureZero();
+
+    // Array Udf defines, udfs either accept array as parameter or returns array
+    void InitArrayUdfs();
 };
 
 }  // namespace udf
